--- lang/jvm/src/main/scala/com/wavesplatform/lang/WavesContext.scala
@@ -1,3 +1,0 @@
-package com.wavesplatform.lang
-
-abstract class WavesContext extends BaseWavesContext
